+++ app/src/main/java/com/example/fse_project/domain/repository/ReservationRepository.kt
@@ -7,11 +7,11 @@ import kotlinx.coroutines.flow.Flow
 interface ReservationRepository {
 
     suspend fun createReservation(reservation: Reservation): Long
-    suspend fun deleteReservation(reservationId: Int)
-    suspend fun updateReservationStatus(reservationId: Int, newStatus: ReservationStatus)
-    suspend fun getReservationById(reservationId: Int): Reservation
+    suspend fun deleteReservation(reservationId: Long)
+    suspend fun updateReservationStatus(reservationId: Long, newStatus: ReservationStatus)
+    suspend fun getReservationById(reservationId: Long): Reservation
     fun getAllReservations(): Flow<List<Reservation>>
-    fun getAllReservationsByUserId(userId: Int): Flow<List<Reservation>>
-    fun getAllReservationsByVehicleId(vehicleId: Int): Flow<List<Reservation>>
-    fun getAllReservationsByChargerId(chargerId: Int): Flow<List<Reservation>>
+    fun getAllReservationsByUserId(userId: Long): Flow<List<Reservation>>
+    fun getAllReservationsByVehicleId(vehicleId: Long): Flow<List<Reservation>>
+    fun getAllReservationsByChargerId(chargerId: Long): Flow<List<Reservation>>
 }
